+++ tests/integration/test_sharepoint.py
@@ -1,17 +1,16 @@
 import os
 import re
+from copy import deepcopy
 
 import pandas as pd
-from copy import deepcopy
 import pytest
 from prefect.tasks.secrets import PrefectSecret
 
 from viadot.config import local_config
 from viadot.exceptions import CredentialError
-from viadot.sources import Sharepoint
+from viadot.sources import Sharepoint, SharepointList
 from viadot.task_utils import df_get_data_types_task
 from viadot.tasks.sharepoint import SharepointToDF
-from viadot.sources import SharepointList
 
 
 def get_url() -> str:
@@ -168,10 +167,11 @@ def test_get_data_types(file_name):
     assert "String" in dtypes
 
 
+### SECTION FOR TESTING SHAREPOINT LIST CONNECTOR ###
 @pytest.fixture(scope="session")
 def sharepoint_list():
     """
-    Fixture for creating a Sharepoint class instance.
+    Fixture for creating a SharepointList class instance.
     The class instance can be used within a test functions to interact with Sharepoint.
     """
     spl = SharepointList()
@@ -187,15 +187,31 @@ def test_valid_filters(sharepoint_list):
     assert result is True
 
 
-def test_invalid_dtype(sharepoint_list):
+def test_filters_missing_dtype(sharepoint_list):
     filters = {
-        "filter1": {"dtype": "list", "operator1": ">", "value1": 10},
+        "filter1": {"operator1": ">", "value1": 10},
     }
-    with pytest.raises(ValueError, match="dtype not allowed!"):
+    with pytest.raises(
+        ValueError,
+        match=re.escape("dtype for filter1 is missing!"),
+    ):
         sharepoint_list.check_filters(filters)
 
 
-def test_missing_operator1(sharepoint_list):
+def test_filters_invalid_dtype(sharepoint_list):
+    filters = {
+        "filter1": {"dtype": "list", "operator1": ">", "value1": 10},
+    }
+    with pytest.raises(
+        ValueError,
+        match=re.escape(
+            "dtype not allowed! Expected: ['datetime', 'date', 'bool', 'int', 'float', 'complex', 'str'] got: list ."
+        ),
+    ):
+        sharepoint_list.check_filters(filters)
+
+
+def test_filters_missing_operator1(sharepoint_list):
     filters = {
         "filter1": {"dtype": "int", "value1": 10},
     }
@@ -203,23 +219,28 @@ def test_missing_operator1(sharepoint_list):
         sharepoint_list.check_filters(filters)
 
 
-def test_invalid_operator1(sharepoint_list):
+def test_filters_invalid_operator1(sharepoint_list):
     filters = {
         "filter1": {"dtype": "int", "operator1": "*", "value1": 10},
     }
-    with pytest.raises(ValueError, match="Operator type not allowed!"):
+    with pytest.raises(
+        ValueError,
+        match=re.escape(
+            "Operator1 type not allowed! Expected: ['<', '>', '<=', '>=', '==', '!='] got: * ."
+        ),
+    ):
         sharepoint_list.check_filters(filters)
 
 
-def test_missing_value1(sharepoint_list):
+def test_filters_missing_value1(sharepoint_list):
     filters = {
         "filter1": {"dtype": "int", "operator1": ">", "value1": None},
     }
-    with pytest.raises(ValueError, match="Value for operator1 is missing!"):
+    with pytest.raises(ValueError, match="Value1 for operator1 is missing!"):
         sharepoint_list.check_filters(filters)
 
 
-def test_missing_operators_conjuction(sharepoint_list):
+def test_filters_missing_operators_conjunction(sharepoint_list):
     filters = {
         "filter1": {
             "dtype": "int",
@@ -227,38 +248,73 @@ def test_missing_operators_conjuction(sharepoint_list):
             "value1": 10,
             "operator2": "<",
             "value2": 20,
-        },
-    }
-    with pytest.raises(ValueError, match="Operators for conjuction is missing!"):
-        sharepoint_list.check_filters(filters)
-
-
-def test_invalid_operators_conjuction(sharepoint_list):
-    filters = {
-        "filter1": {
-            "dtype": "int",
-            "operator1": ">",
-            "value1": 10,
-            "operator2": "<",
-            "value2": 20,
-            "operators_conjuction": "!",
-        },
-    }
-    with pytest.raises(ValueError, match="Operators for conjuction not allowed!"):
-        sharepoint_list.check_filters(filters)
-
-
-def test_invalid_filters_conjuction(sharepoint_list):
-    filters = {
-        "filter1": {
-            "dtype": "int",
-            "operator1": ">",
-            "value1": 10,
-            "filters_conjuction": "!",
         },
     }
     with pytest.raises(
-        ValueError, match="Filters operators for conjuction not allowed!"
+        ValueError,
+        match=re.escape(
+            "Operator for conjunction is missing! Expected: ['&', '|'] got empty."
+        ),
+    ):
+        sharepoint_list.check_filters(filters)
+
+
+def test_filters_invalid_operators_conjunction(sharepoint_list):
+    filters = {
+        "filter1": {
+            "dtype": "int",
+            "operator1": ">",
+            "value1": 10,
+            "operator2": "<",
+            "value2": 20,
+            "operators_conjunction": "!",
+        },
+    }
+    with pytest.raises(
+        ValueError,
+        match=re.escape(
+            "Operator for conjunction not allowed! Expected: ['&', '|'] got ! ."
+        ),
+    ):
+        sharepoint_list.check_filters(filters)
+
+
+def test_filters_conjunction_not_allowed(sharepoint_list):
+    filters = {
+        "filter1": {
+            "dtype": "int",
+            "operator1": ">",
+            "value1": 10,
+            "filters_conjunction": "!",
+        },
+    }
+    with pytest.raises(
+        ValueError,
+        match=re.escape(
+            "Filters conjunction allowed only when more then one filter provided!"
+        ),
+    ):
+        sharepoint_list.check_filters(filters)
+
+
+def test_filters_invalid_conjunction(sharepoint_list):
+    filters = {
+        "filter1": {
+            "dtype": "int",
+            "value1": 10,
+            "operator1": ">",
+            "filters_conjunction": "!",
+        },
+        "filter2": {
+            "dtype": "int",
+            "operator1": "==",
+        },
+    }
+    with pytest.raises(
+        ValueError,
+        match=re.escape(
+            "Filter operator for conjunction not allowed! Expected: ['&', '|'] got ! ."
+        ),
     ):
         sharepoint_list.check_filters(filters)
 
@@ -266,52 +322,86 @@ def test_invalid_filters_conjuction(sharepoint_list):
 def test_valid_mapping(sharepoint_list):
     filters = {
         "filter1": {
+            "dtype": "int",
+            "value1": 10,
+            "value2": 20,
             "operator1": ">",
             "operator2": "<=",
-            "operators_conjuction": "&",
-            "filters_conjuction": "|",
+            "operators_conjunction": "&",
+            "filters_conjunction": "|",
         },
-        "filter2": {"operator1": "==", "operator2": "!=", "operators_conjuction": "|"},
+        "filter2": {
+            "dtype": "int",
+            "value1": 30,
+            "value2": 0,
+            "operator1": "==",
+            "operator2": "!=",
+            "operators_conjunction": "|",
+        },
     }
     expected_result = {
         "filter1": {
+            "dtype": "int",
+            "value1": 10,
+            "value2": 20,
             "operator1": "gt",
             "operator2": "le",
-            "operators_conjuction": "and",
-            "filters_conjuction": "or",
+            "operators_conjunction": "and",
+            "filters_conjunction": "or",
         },
-        "filter2": {"operator1": "eq", "operator2": "ne", "operators_conjuction": "or"},
+        "filter2": {
+            "dtype": "int",
+            "value1": 30,
+            "value2": 0,
+            "operator1": "eq",
+            "operator2": "ne",
+            "operators_conjunction": "or",
+        },
     }
-    result = sharepoint_list.operators_mapping(deepcopy(filters))
+    result = sharepoint_list.operators_mapping(filters)
     assert result == expected_result
 
 
-def test_invalid_comparison_operator(sharepoint_list):
+def test_operators_mapping_invalid_comparison_operator(sharepoint_list):
     filters = {
         "filter1": {
             "operator1": "*",
             "operator2": "<=",
-            "operators_conjuction": "&",
-            "filters_conjuction": "|",
+            "operators_conjunction": "&",
+            "filters_conjunction": "|",
         },
     }
     error_message = "This comparison operator: * is not allowed. Please read the function documentation for details!"
     with pytest.raises(ValueError, match=re.escape(error_message)):
-        sharepoint_list.operators_mapping(deepcopy(filters))
+        sharepoint_list.operators_mapping(filters)
 
 
-def test_invalid_logical_operator(sharepoint_list):
+def test_operators_mapping_invalid_logical_operator(sharepoint_list):
     filters = {
         "filter1": {
             "operator1": ">",
             "operator2": "<=",
-            "operators_conjuction": "!",
-            "filters_conjuction": "|",
+            "operators_conjunction": "!",
+            "filters_conjunction": "|",
         },
     }
-    error_message = "This conjuction(logical) operator: ! is not allowed. Please read the function documentation for details!"
+    error_message = "This conjunction (logical) operator: ! is not allowed. Please read the function documentation for details!"
     with pytest.raises(ValueError, match=re.escape(error_message)):
-        sharepoint_list.operators_mapping(deepcopy(filters))
+        sharepoint_list.operators_mapping(filters)
+
+
+def test_operators_mapping_invalid_filters_logical_operator(sharepoint_list):
+    filters = {
+        "filter1": {
+            "operator1": ">",
+            "operator2": "<=",
+            "operators_conjunction": "&",
+            "filters_conjunction": "!",
+        },
+    }
+    error_message = "This filters conjunction (logical) operator: ! is not allowed. Please read the function documentation for details!"
+    with pytest.raises(ValueError, match=re.escape(error_message)):
+        sharepoint_list.operators_mapping(filters)
 
 
 def test_single_filter_datetime_api(sharepoint_list):
@@ -348,7 +438,7 @@ def test_single_df_filter(sharepoint_list):
 
 def test_multiple_df_filters(sharepoint_list):
     filters = {
-        "column1": {"operator1": ">", "value1": 10, "filters_conjuction": "&"},
+        "column1": {"operator1": ">", "value1": 10, "filters_conjunction": "&"},
         "column2": {"operator1": "<", "value1": 20},
     }
     result = sharepoint_list.make_filter_for_df(filters)