+++ viadot/flows/sharepoint_to_adls.py
@@ -17,8 +17,7 @@ from viadot.task_utils import (
     validate_df,
 )
 from viadot.tasks import AzureDataLakeUpload
-from viadot.tasks.sharepoint import SharepointToDF, SharepointListToDF
-
+from viadot.tasks.sharepoint import SharepointListToDF, SharepointToDF
 
 logger = logging.get_logger()
 
@@ -42,6 +41,7 @@ class SharepointToADLS(Flow):
         if_exists: str = "replace",
         validate_df_dict: dict = None,
         timeout: int = 3600,
+        set_prefect_kv: bool = False,
         *args: List[any],
         **kwargs: Dict[str, Any],
     ):
@@ -69,6 +69,7 @@ class SharepointToADLS(Flow):
             dataframe. If defined, triggers the `validate_df` task from task_utils. Defaults to None.
             timeout(int, optional): The amount of time (in seconds) to wait while running this task before
                 a timeout occurs. Defaults to 3600.
+            set_prefect_kv (bool, optional): Whether to do key-value parameters in KV Store or not. Defaults to False.
         """
         # SharepointToDF
         self.if_empty = if_empty
@@ -86,6 +87,7 @@ class SharepointToADLS(Flow):
         self.adls_sp_credentials_secret = adls_sp_credentials_secret
         self.if_exists = if_exists
         self.output_file_extension = output_file_extension
+        self.set_prefect_kv = set_prefect_kv
         self.now = str(pendulum.now("utc"))
         if self.local_dir_path is not None:
             self.local_file_path = (
@@ -177,7 +179,8 @@ class SharepointToADLS(Flow):
 
         file_to_adls_task.set_upstream(df_to_file, flow=self)
         json_to_adls_task.set_upstream(dtypes_to_json_task, flow=self)
-        set_key_value(key=self.adls_dir_path, value=self.adls_file_path)
+        if self.set_prefect_kv == True:
+            set_key_value(key=self.adls_dir_path, value=self.adls_file_path)
 
     @staticmethod
     def slugify(name):
@@ -188,39 +191,40 @@ class SharepointListToADLS(Flow):
     def __init__(
         self,
         name: str,
-        list_title: str = None,
-        site_url: str = None,
+        list_title: str,
+        site_url: str,
+        path: str,
+        adls_dir_path: str,
+        adls_file_name: str,
+        filters: dict = None,
         required_fields: List[str] = None,
         field_property: str = "Title",
-        filters: dict = None,
         row_count: int = 5000,
+        adls_sp_credentials_secret: str = None,
         sp_cert_credentials_secret: str = None,
         vault_name: str = None,
-        path: str = None,
-        adls_dir_path: str = None,
-        adls_file_name: str = None,
-        adls_sp_credentials_secret: str = None,
         overwrite_adls: bool = True,
         output_file_extension: str = ".parquet",
         validate_df_dict: dict = None,
+        set_prefect_kv: bool = False,
         *args: List[any],
         **kwargs: Dict[str, Any],
     ):
         """
-        Run Flow SharepointListToADLS.
+        Flow for ingesting sharepoint list items(rows) with a given (or all) columns.
+        It allows to filter the output by column values.
+        Data is ingested from MS Sharepoint list (with given name and url ) and stored in MS Azure ADLS.
 
         Args:
-        name (str): Prefect flow name.
-        list_title (str): Title of Sharepoint List. Default to None.
-        site_url (str): URL to set of Sharepoint Lists. Default to None.
-        required_fields (List[str]): Required fields(columns) need to be extracted from
-                                     Sharepoint List. Default to None.
-        field_property (List[str]): Property to expand with expand query method.
-                                    All propertys can be found under list.item.properties.
-                                    Default to ["Title"]
-        filters (dict): Dictionary with operators which filters the SharepointList output.
+            name (str): Prefect flow name.
+            list_title (str): Title of Sharepoint List.
+            site_url (str): URL to set of Sharepoint Lists.
+            path (str): Local file path. Default to None.
+            adls_dir_path (str): Azure Data Lake destination folder/catalog path. Defaults to None.
+            adls_file_name (str): Name of file in ADLS. Defaults to None.
+            filters (dict, optional): Dictionary with operators which filters the SharepointList output. Defaults to None.
                         allowed dtypes: ('datetime','date','bool','int', 'float', 'complex', 'str')
-                        allowed conjuction: ('&','|')
+                        allowed conjunction: ('&','|')
                         allowed operators: ('<','>','<=','>=','==','!=')
                         Example how to build the dict:
                         filters = {
@@ -231,8 +235,8 @@ class SharepointListToADLS(Flow):
                                 'value2':'YYYY-MM-DD',
                                 'operator1':'>=',
                                 'operator2':'<=',
-                                'operators_conjuction':'&', # conjuction operators allowed only when 2 values passed
-                                'filters_conjuction':'&', # conjuction filters allowed only when 2 columns passed
+                                'operators_conjunction':'&', # conjunction operators allowed only when 2 values passed
+                                'filters_conjunction':'&', # conjunction filters allowed only when 2 columns passed
                                 }
                                 ,
                         'Column_name_2' :
@@ -242,16 +246,26 @@ class SharepointListToADLS(Flow):
                                 'operator1':'==',
                                 },
                         }
-        row_count (int): Number of downloaded rows in single request. Default to 5000.
-        sp_cert_credentials_secret (str): Credentials to verify Sharepoint connection. Default to None.
-        vault_name (str): KeyVaultSecret name. Default to None.
-        path (str): Local file path. Default to None.
-        adls_dir_path (str): Azure Data Lake destination folder/catalog path. Defaults to None.
-        adls_file_name (str, optional): Name of file in ADLS. Defaults to None.
-        adls_sp_credentials_secret (str, optional): The name of the Azure Key Vault secret containing a dictionary with
-                                                    ACCOUNT_NAME and Service Principal credentials (TENANT_ID, CLIENT_ID,
-                                                    CLIENT_SECRET) for the Azure Data Lake. Defaults to None.
-        overwrite_adls (bool, optional): Whether to overwrite files in the lake. Defaults to True.
+            required_fields (List[str], optional): Required fields(columns) need to be extracted from
+                                     Sharepoint List. Defaults to None.
+            field_property (str, optional): Property to expand fields with expand query method.
+                                    For example: User fields could be expanded and "Title"
+                                    or "ID" could be extracted
+                                    -> useful to get user name instead of ID
+                                    All properties can be found under list.item.properties.
+                                    WARNING! Field types and properties might change which could
+                                    lead to errors - extension of sp connector would be required.
+                                    Default to ["Title"]. Defaults to "Title".
+            row_count (int, optional): Number of downloaded rows in single request.Defaults to 5000.
+            adls_sp_credentials_secret (str, optional): Credentials to connect to Azure ADLS
+                                    If not passed it will take cred's from your .config/credentials.json Defaults to None.
+            sp_cert_credentials_secret (str, optional): Credentials to verify Sharepoint connection.
+                                    If not passed it will take cred's from your .config/credentials.json Default to None.
+            vault_name (str, optional): KeyVaultSecret name. Default to None.
+            overwrite_adls (bool, optional): Whether to overwrite files in the lake. Defaults to True.
+            output_file_extension (str, optional): Extension of the resulting file to be stored. Defaults to ".parquet".
+            validate_df_dict (dict, optional): Whether to do an extra df validation before ADLS upload or not to do. Defaults to None.
+            set_prefect_kv (bool, optional): Whether to do key-value parameters in KV Store or not. Defaults to False.
 
         Returns:
             .parquet file inside ADLS.
@@ -275,6 +289,7 @@ class SharepointListToADLS(Flow):
         self.overwrite = overwrite_adls
         self.adls_sp_credentials_secret = adls_sp_credentials_secret
         self.output_file_extension = output_file_extension
+        self.set_prefect_kv = set_prefect_kv
         self.now = str(pendulum.now("utc"))
         if self.path is not None:
             self.local_file_path = (
@@ -365,7 +380,8 @@ class SharepointListToADLS(Flow):
 
         file_to_adls_task.set_upstream(df_to_file, flow=self)
         json_to_adls_task.set_upstream(dtypes_to_json_task, flow=self)
-        set_key_value(key=self.adls_dir_path, value=self.adls_file_path)
+        if self.set_prefect_kv == True:
+            set_key_value(key=self.adls_dir_path, value=self.adls_file_path)
 
     @staticmethod
     def slugify(name):