+++ viadot/sources/sap_bw.py
@@ -101,7 +101,7 @@ class SAPBW(Source):
                     {
                     "COLUMN": 0,
                     "ROW": 0,
-                    "DATA": "VELUX Deutschland GmbH",
+                    "DATA": "DATA",
                     "VALUE_DATA_TYPE": "CHAR",
                     "CELL_STATUS": ""
                     },...